Add static assertions to ensure RefCounted and ThreadSafeRefCounted are the same size https://bugs.webkit.org/show_bug.cgi?id=271337
Reviewed by NOBODY (OOPS!).

Currently GeneratedSerializers.cpp assumes that RefCounted and
ThreadSafeRefCounted are both the same size. Arguably it shouldn't,
since this surely relies on implementation-defined behavior, but it
apparently works in practice. If it's not true, then the build will fail
on one of various static assertions in GeneratedSerializers.cpp, because
VirtualTableAndRefCountOverhead inherits from
RefCounted<VirtualTableAndRefCountOverhead> and gets compared to the
size of ThreadSafeRefCounted objects. Since it's easy to test for, let's
add more static asserts to catch this problem specifically. This way, if
the assumption fails in the future, the sad developer who has to deal
with it will hopefully be moderately less confused.

It would be better to check against the size of ThreadSafeRefCounted
instead of assuming that it's the same size as RefCounted, but that
would require more work.

Let's also adjust the preprocessor guards in ThreadSafeRefCounted.h to
match RefCounted.h. The size of both classes depends on whether debug
assertions are enabled, but ThreadSafeRefCounted is directly checking
NDEBUG rather than checking ASSERT_ENABLED, so they could easily wind up
enabled in one place and not the other. This would likely cause the size
checks in GeneratedSerializers.cpp to fail.

* Source/WTF/wtf/ThreadSafeRefCounted.h:
* Source/WebKit/Scripts/generate-serializers.py:
(generate_impl):