[web-animations] add support for canceling style-originated animations silently https://bugs.webkit.org/show_bug.cgi?id=271365
Reviewed by Antti Koivisto.

When we will be adding animation support for the `display` property (see bug 267762) we will
need to resolve animations even when the underlying style has `display: none` to see whether
animations yield a different value. In that process, we may create style-originated animations
that will ultimately be canceled once we're certain `display: none` is still set.

In 276414@main we created a list of style-originated animations created during a style update.

We now add support for canceling that group of style-originated animations in a way that is not
script observable, or "silently" in Web Animations parlance.

Once we get to add animation support for the `display` property proper, we will add a call to
`StyleOriginatedAnimation::cancelFromStyle()` with the list of newly-created style-originated
animations that require silent cancelation.

* Source/WebCore/animation/StyleOriginatedAnimation.cpp:
(WebCore::StyleOriginatedAnimation::cancel):
(WebCore::StyleOriginatedAnimation::cancelFromStyle):
* Source/WebCore/animation/StyleOriginatedAnimation.h:
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::cancel):
* Source/WebCore/animation/WebAnimation.h:
* Source/WebCore/style/Styleable.cpp:
(WebCore::Styleable::cancelStyleOriginatedAnimations const):
* Source/WebCore/style/Styleable.h:
(WebCore::Styleable::cancelStyleOriginatedAnimations):

Canonical link: https://commits.webkit.org/276453@main