`runInCrossOriginFrame` should use localhost as the mainframe URL https://bugs.webkit.org/show_bug.cgi?id=271414 rdar://125193552
Reviewed by Alex Christensen.

Some tests expect that the mainframe is hosted locally, but it will currently not be with
`runInCrossOriginFrame=true`. We can make the mainframe hosted locally if we use localhost instead of
webkit.org. The iframe URL will be http://127.0.0.1/, so it will still be cross-origin.

* LayoutTests/platform/mac-site-isolation/TestExpectations:
* Tools/WebKitTestRunner/TestInvocation.cpp:
(WTR::TestInvocation::invoke):

Canonical link: https://commits.webkit.org/276501@main