REGRESSION (276015@main): [iOS] WKHTTPCookieStore.SameSiteWithPatternMatch is a constant failure https://bugs.webkit.org/show_bug.cgi?id=271398 rdar://125179058
Reviewed by Alex Christensen.

The test that I added in 276015@main is buggy. It passes on macOS for strange
and racy reasons, the iOS failure is correct. This change: 1) relaxes
third-party cookie blocking, which is necessary for testing the intended
behavior, 2) waits until a load finishes before proceeding with some tests, and
3) fixes a typo.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm:
(TEST):

Canonical link: https://commits.webkit.org/276529@main