Web Inspector: Timelines: DataGrid header layout issues in some locales https://bugs.webkit.org/show_bug.cgi?id=271481 rdar://124993083
Reviewed by NOBODY (OOPS!).

The line box of text is defined by the font used to render glyphs.

- When using `overflow: hidden`, the clipping region is given by the height of the line box,
not the glyphs in it. With an insufficient line height, text can get clipped.

Fixed by defining a line-height value.

- When the line box is taller, it increases the dimensions of the element.
The DataGrid header was positioned in the document flow.
The DataGrid content is absolutely positioned (out-of-flow) and offset vertically by an amount
equal to the height of the the header. When the header grows, it goes underneath the content.

Fixed by placing the DataGrid header on top of the content.
Overlap might still occur in some scenarios, but it will obscure a top portion
of the first row of content instead of hiding the header underneath it.

Ideally the layout of DataGrid would not be based on absolute positioning
and fixed dimensions, but that refactoring is beyond the scope of this patch.

* Source/WebInspectorUI/UserInterface/Views/DataGrid.css:
(.data-grid):
(.data-grid .data-container):
(.data-grid.inline):
(.data-grid > .header-wrapper):
(.data-grid th > .header-cell-content):