AX: Mac combo boxes should expose aria-activedescendant as AXSelectedChildren https://bugs.webkit.org/show_bug.cgi?id=271287 rdar://problem/125050867
Reviewed by Andres Gonzalez.

This is necessary, but probably not sufficient, to support
https://bugs.webkit.org/show_bug.cgi?id=231724 (AX: VoiceOver doesn't
announce aria-activedescendant value for select-only combobox
specification). There are other bugs preventing VoiceOver from working
but exposing the active descendant is pretty fundamental so I'd like
to get that in first.

Part of the bug is that an element can be both a combo box and a text
field, so the supported attribute logic needed to be updated to take
that into account.

* LayoutTests/accessibility/mac/aria-combobox-activedescendant-expected.txt: Added.
* LayoutTests/accessibility/mac/aria-combobox-activedescendant.html: Added.
* Source/WebCore/accessibility/AXCoreObject.cpp:
(WebCore::AXCoreObject::canHaveSelectedChildren const):
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::selectedChildren):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper ALLOW_DEPRECATED_IMPLEMENTATIONS_END]):

Canonical link: https://commits.webkit.org/276627@main