[WebGPU] Partial 3D image_copy tests are failing https://bugs.webkit.org/show_bug.cgi?id=271382 <radar://125168133>
Reviewed by Tadeu Zagallo.

For the purpose of validation of bytesPerRow, we should be
using the 2D dimension as the issue is we would crash at the metal validation layer.

No test expectations update since the expectations were set to passing and this
test does not run in EWS as it takes too long.

* Source/WebGPU/WebGPU/CommandEncoder.mm:
(WebGPU::CommandEncoder::copyBufferToTexture):
(WebGPU::CommandEncoder::copyTextureToBuffer):
* Source/WebGPU/WebGPU/Queue.mm:
(WebGPU::Queue::writeTexture):

Canonical link: https://commits.webkit.org/276649@main