AX: [WKAccessibilityWebPageObject accessibilityAttributeValue:] should retrieve several properties from the main thread. https://bugs.webkit.org/show_bug.cgi?id=271675 <rdar://problem/125374335>
Reviewed by Tyler Wilcock.

Several of the AX API attributes including NSAccessibilityWindowAttribute and NSAccessibilityPrimaryScreenHeightAttribute need to be retrieved from the main thread. Among other problems, this may be the cause of intermittent crashes we observe in accessibility/primary-screen-height.html. This patch changes this method to execute on the AX thread only those properties that are safe off the main thread, and to dispatch to the main thread the rest.

* Source/WebKit/WebProcess/WebPage/mac/WKAccessibilityWebPageObjectMac.mm:
(-[WKAccessibilityWebPageObject accessibilityAttributeValue:]):

Canonical link: https://commits.webkit.org/276686@main