check-webkit-style should warn when using soft-linking https://bugs.webkit.org/show_bug.cgi?id=271270 <rdar://125038029>
Reviewed by NOBODY (OOPS!).

The goal of this warning is to have a discussion on Slack about the
specific use case each time a new framework or library is soft-linked,
in case there is an alternative.  Adding more soft-linked frameworks or
libraries will become more work in the future, and may result in build
failures if not done correctly.

* Tools/Scripts/webkitpy/style/checkers/cpp.py:
(check_language):
- Separate check for SOFT_LINK macros in header files from source files.
- Always emit a warning when using SOFT_LINK macros for frameworks or
  libraries now.
(CppChecker):
- Add new 'softlink/dynamic_linking' category.
* Tools/Scripts/webkitpy/style/checkers/cpp_unittest.py:
(CppStyleTest.test_softlink_framework):
- Remove test for SOFT_LINK_FRAMEWORK_OPTIONAL_PREFLIGHT() macro which
  doesn't exist anymore.
- Update test results for new warning.
- Add tests for SOFT_LINK framework macros that were missing.
(CppStyleTest.test_softlink_library): Add.
- Add tests for SOFT_LINK library macros.