[iOS] AVVideoCaptureSource is not handling the case of another application changing the device videoZoomFactor https://bugs.webkit.org/show_bug.cgi?id=271643 rdar://125351159
Reviewed by Eric Carlson.

When another application is changing the device zoom, it got access to the device so AVVideoCaptureSource is not running.
When AVVideoCaptureSource restarts, we just have to reapply the device zoom.

Remove AVVideoCaptureSource::m_appliedZoom since the device videoZoomFactor can change when other iOS apps get access to the camera.
We thus always compare the zoom to be applied with the device zoom.

* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.h:
* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.mm:
(WebCore::AVVideoCaptureSource::setSessionSizeFrameRateAndZoom):

Canonical link: https://commits.webkit.org/276731@main