Camera is paused occasionally when torch is enabled https://bugs.webkit.org/show_bug.cgi?id=270831 rdar://124434403
Reviewed by Andy Estes.

We do not want to update AV capture devices before the session is running, as it may trigger interruption.
We now apply changes to device as soon as the session is running, but not before.
We apply this to torch, white balance, resolution and frame rate.

Whenever we want to update any of these, if the session is not running, we exit early and we remember to reconfigure.
When the session resumes, we check whether we need to reconfigure and do so.

* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.h:
* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.mm:
(WebCore::AVVideoCaptureSource::setSessionSizeFrameRateAndZoom):
(WebCore::AVVideoCaptureSource::updateWhiteBalanceMode):
(WebCore::AVVideoCaptureSource::updateTorch):
(WebCore::AVVideoCaptureSource::reconfigure):
(WebCore::AVVideoCaptureSource::captureSessionIsRunningDidChange):

Canonical link: https://commits.webkit.org/276773@main