[iOS] Allow camera zoom above 2 https://bugs.webkit.org/show_bug.cgi?id=271638 rdar://125180182
Reviewed by NOBODY (OOPS!).

Zoom is expected to be bound to elevated PTZ permission according the image capture specification.
But this is more related to pan and tilt than zoom.
We increase the zoom from 2 (it was expected to be 4) to 10 without using any elevated PTZ permission.
This should cover most needs.

* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.mm:
(WebCore::AVVideoCaptureSource::computeMaxZoom const):