AX: The selected children property of a list that is not owned or controlled by its combobox is not updated in ITM. https://bugs.webkit.org/show_bug.cgi?id=271783 <rdar://problem/125504735>
Reviewed by Tyler Wilcock.

Covered by accessibility/aria-combobox-no-owns.html, that now passes in ITM.

When the combobox does not own or control its list, we were updating the combobox object but not the list. We now check for this scenario and update the list's selected children as appropriate.

* LayoutTests/accessibility-isolated-tree/TestExpectations:
* LayoutTests/accessibility/aria-combobox-no-owns-expected.txt:
* LayoutTests/accessibility/aria-combobox-no-owns.html:
Modernized the code of this test.

* Source/WebCore/accessibility/AXCoreObject.h:
(WebCore::Accessibility::findRelatedObjectInAncestry):
Renamed variable in the for loop for readability, to avoid confusion with the parameter that had the same name.

* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::handleActiveDescendantChange):

Canonical link: https://commits.webkit.org/276779@main