[visionOS] Unable to set playback speed back to 1x after changing it to a different speed https://bugs.webkit.org/show_bug.cgi?id=271813 rdar://125522760
Reviewed by Eric Carlson.

In LinearMediaPlayer we set selectedPlaybackRate to 1 and never updated it when the playback rate
changed. As a result, if you set the playback rate to, e.g., 1.25x then attempt to change it back to
1x, LinearMediaKit would never call playbackRateChanged() on our Playable object because it belives
1x is already the selected rate.

Resolved this by updating selectedPlaybackRate when the playback rate changes.

* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
(WebKit::PlaybackSessionInterfaceLMK::rateChanged):

Canonical link: https://commits.webkit.org/276795@main